moved ConnectToWallet into core pkg
