+++ Packages/io.chainsafe.web3-unity/Runtime/Scripts/ConnectToWallet.cs.meta
