--- Packages/io.chainsafe.web3-unity/Runtime/Scripts/Samples/MultiCallSample.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1724aa460bb84aa3aececd9d7cc49c78
-timeCreated: 1696430378
